# Speed up walkLength for narrow strings